FIX: app/page.tsx를 서버 컴포넌트로 분리 - 기존 달력/투두리스트 컴포넌트를 MainClient로 분리
